backend/requirements.txt
# =============================================================================
# AutoVNC Backend Dependencies
# =============================================================================

# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# WebSocket support
websockets==12.0

# VNC client
vncdotool==1.2.0

# Computer vision
opencv-python-headless==4.9.0.80
numpy==1.26.4

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# HTTP client for AI API calls
httpx==0.27.0
openai==1.12.0

# Data validation
pydantic==2.6.1
pydantic-settings==2.2.1

# Async utilities
aiofiles==23.2.1

# Docker SDK for runner management
docker==7.0.0

# Utilities
python-dotenv==1.0.1
